[stats_service] split tests by RPC verb; lift errors to finelog.errors
Split the two oversized stats-service test files into focused per-verb
files and break the circular import they were papering over.

- test_stats_service.py + test_stats_query_drop.py (1432 lines, 51 tests)
  → conftest.py + test_register.py + test_write_rows.py + test_query.py
  + test_drop.py + test_eviction.py + test_concurrency.py
  + test_persistence.py + test_asgi.py. All ASGI/Connect tests now live
  together in test_asgi.py instead of being scattered. Test count
  unchanged (144 collected before/after).

- Move finelog.client.errors → finelog.errors (top of package). The old
  module remains as a thin re-export for backwards compatibility.
  finelog.store.schema now imports from the new location, breaking the
  cycle where importing DuckDBLogStore in isolation triggered
  schema → finelog.client.__init__ → log_client → schema (still loading).
